move service tests to service folder
